reorganize frontend files, break down into separated components
