+++ .gitignore
@@ -3,6 +3,7 @@
 .vscode/
 
 __pycache__/
+.pytest_cache/
 *.pyc
 *.pyo
 *.pyd
