+++ frontend/index.html
@@ -10,7 +10,8 @@
   <link rel="icon" type="image/png" href="/static/android-chrome-192x192.png" sizes="192x192">
   <link rel="icon" type="image/png" href="/static/android-chrome-512x512.png" sizes="512x512">
   <link rel="apple-touch-icon" href="/static/apple-touch-icon.png" />
-  <link href="https://fonts.googleapis.com/css2?family=Roboto+Mono:wght@700&family=Clear+Sans:wght@500&display=swap" rel="stylesheet">
+  <link href="https://fonts.googleapis.com/css2?family=Roboto+Mono:wght@700&family=Clear+Sans:wght@500&display=swap"
+    rel="stylesheet">
   <title>Osmosmjerka</title>
 </head>
