+++ frontend/src/components/AdminPanel/AdminPanel.css
